refactor(web): split domain/assignments.ts into operation sub-modules behind a barrel
Decompose the 2443-line assignments domain monolith into cohesive
./assignments/* sub-modules (createEdit, repoCreation, copyReuse, autograderYaml,
deleteAssignment, permissions, accept) over a leaf accessPrimitives.ts holding
the shared template-access + accept-step scaffolding (parseTemplateRef,
resolveTemplate, verifyTemplateAccess, withAcceptStep + step types,
contentsPathExists, resolveConfigRepoDefaultBranch, scoped log). accept.ts is the
heavy consumer, pulling template/yaml/permissions/accept-step downward.

preserveUnmanagedAssignmentKeys stays in createEdit.ts (its sole value-consumer)
to avoid a createEdit<->copyReuse value cycle; the remaining back-edges are
type-only, which the no-cycle guard ignores. withGitConflictRetry stays in
domain/classrooms.ts; isOwnerGitHubOrgRole keeps its @/authz barrel import.

assignments.ts is now a thin re-export barrel: all 27 exports preserved, so the
~8 importers are unchanged. Behavior-preserving — tsc + full vitest (1533) green,
no-cycle + boundaries + arch:validate green; assignments.test.ts still exercises
the public surface via the barrel. Completes Tier-2 Phase D (all 4 data-layer
monoliths decomposed).
