+++ web/src/domain/assignments/deleteAssignment.ts
@@ -1,0 +1,94 @@
+import type { GitHubClient } from "@/github-core/client"
+import {
+  getBranchRef,
+  getCommit,
+  getConfigRepoBranch,
+} from "@/github-core/configRepoReads"
+import { prefixCommit } from "@/util/commit"
+import {
+  createGitCommit,
+  createGitTree,
+  updateRef,
+} from "@/github-core/mutations"
+import { getAssignmentsFile } from "../queries/assignments"
+import { assertClassroomNotArchived } from "../classrooms"
+import { log } from "./accessPrimitives"
+
+export type DeleteAssignmentInput = {
+  org: string
+  classroom: string
+  assignment: string
+}
+export async function deleteAssignment(
+  client: GitHubClient,
+  input: DeleteAssignmentInput,
+) {
+  const { org, classroom, assignment: slug } = input
+
+  log.info("delete assignment: started", { org, classroom, slug })
+
+  // Refuse a delete into an archived classroom (write-path guard); run the
+  // check concurrently with the ref read.
+  const [, configBranch] = await Promise.all([
+    assertClassroomNotArchived(client, org, classroom),
+    getConfigRepoBranch(client, org),
+  ])
+  const ref = await getBranchRef(client, org, configBranch)
+  const commit = await getCommit(client, org, ref.object.sha)
+
+  const assignmentsFilePath = `${classroom}/assignments.json`
+  const currentAssignments = await getAssignmentsFile(client, {
+    org,
+    path: assignmentsFilePath,
+    ref: ref.object.sha,
+  })
+
+  const targetAssignment = currentAssignments.assignments.find(
+    (a) => a.slug === slug,
+  )
+
+  if (!targetAssignment) {
+    throw new Error(`Existing assignment matching ${slug} was not found.`)
+  }
+
+  const nextAssignments = {
+    ...currentAssignments,
+    assignments: [
+      ...currentAssignments.assignments.filter((a) => a.slug !== slug),
+    ],
+  }
+
+  const tree = await createGitTree(client, {
+    org: input.org,
+    base_tree: commit.tree.sha,
+    tree: [
+      {
+        path: assignmentsFilePath,
+        mode: "100644",
+        type: "blob",
+        content: JSON.stringify(nextAssignments, null, 2) + "\n",
+      },
+    ],
+  })
+
+  const newCommit = await createGitCommit(client, {
+    org: input.org,
+    message: prefixCommit(`Edit assignment: ${input.classroom}/${slug}`),
+    tree_sha: tree.sha,
+    parents: [ref.object.sha],
+  })
+  const updatedRef = await updateRef(
+    client,
+    input.org,
+    newCommit.sha,
+    configBranch,
+  )
+
+  return {
+    previousCommitSha: ref.object.sha,
+    baseTreeSha: commit.tree.sha,
+    newTreeSha: tree.sha,
+    newCommitSha: newCommit.sha,
+    updatedRef,
+  }
+}
